draft of the search using haystack+elasticsearch
You need to install a local elasticsearch node (elasticsearch.com) and
start it using the default properties. in the app's directory run
manage.py rebuild_index to initialize the index and you should be ready to
go. Try by creating a channel and searching for it (no search box, goto
/search/)